Feature/dracelinux 32

See merge request multicore/drace!102

+++ drace-client/detectors/traceBinary/TraceBinary.cpp
@@ -55,7 +55,7 @@ namespace drace
 
             virtual void func_enter(tls_t tls, void* pc) {
                 ipc::event::BufferEntry buf{Type::FUNCENTER};
-                buf.payload.funcenter = {(uint32_t)(uintptr_t)tls, (uint64_t)pc};
+                buf.payload.funcenter = {(uint32_t)(uintptr_t)tls, (uintptr_t)pc};
                 write_log_sync(buf);
             }
 
@@ -72,7 +72,7 @@ namespace drace
                 bool write)
             {
                 ipc::event::BufferEntry buf{Type::ACQUIRE};
-                buf.payload.mutex = {(uint32_t)(uintptr_t)tls, (uint64_t)mutex, (int)recursive, write, true};
+                buf.payload.mutex = {(uint32_t)(uintptr_t)tls, (uintptr_t)mutex, (int)recursive, write, true};
                 write_log_sync(buf);
             }
 
@@ -81,50 +81,50 @@ namespace drace
                 void* mutex,
                 bool write) {
                     ipc::event::BufferEntry buf{Type::RELEASE};
-                    buf.payload.mutex = {(uint32_t)(uintptr_t)tls, (uint64_t)mutex, (int)0, write, false};
+                    buf.payload.mutex = {(uint32_t)(uintptr_t)tls, (uintptr_t)mutex, (int)0, write, false};
                     write_log_sync(buf);
                 }
 
             virtual void happens_before(tls_t tls, void* identifier) {
                 ipc::event::BufferEntry buf{Type::HAPPENSBEFORE};
-                buf.payload.happens = {(uint32_t)(uintptr_t)tls, (uint64_t)identifier};
+                buf.payload.happens = {(uint32_t)(uintptr_t)tls, (uintptr_t)identifier};
                 write_log_sync(buf);
             }
 
             virtual void happens_after(tls_t tls, void* identifier) {
                 ipc::event::BufferEntry buf{Type::HAPPENSAFTER};
-                buf.payload.happens = {(uint32_t)(uintptr_t)tls, (uint64_t)identifier};
+                buf.payload.happens = {(uint32_t)(uintptr_t)tls, (uintptr_t)identifier};
                 write_log_sync(buf);
             }
 
             virtual void read(tls_t tls, void* pc, void* addr, size_t size) {
                 ipc::event::BufferEntry buf{Type::MEMREAD};
-                buf.payload.memaccess = {(uint32_t)(uintptr_t)tls, (uint64_t)pc, (uint64_t)addr, (uint64_t)size};
+                buf.payload.memaccess = {(uint32_t)(uintptr_t)tls, (uintptr_t)pc, (uintptr_t)addr, (uintptr_t)size};
                 write_log_sync(buf);
             }
 
             virtual void write(tls_t tls, void* pc, void* addr, size_t size) {
                 ipc::event::BufferEntry buf{Type::MEMWRITE};
-                buf.payload.memaccess = {(uint32_t)(uintptr_t)tls, (uint64_t)pc, (uint64_t)addr, (uint64_t)size};
+                buf.payload.memaccess = {(uint32_t)(uintptr_t)tls, (uintptr_t)pc, (uintptr_t)addr, (uintptr_t)size};
                 write_log_sync(buf);
             }
 
             virtual void allocate(tls_t tls, void* pc, void* addr, size_t size) {
                 ipc::event::BufferEntry buf{Type::ALLOCATION};
-                buf.payload.allocation = {(uint32_t)(uintptr_t)tls, (uint64_t)pc, (uint64_t)addr, (uint64_t)size};
+                buf.payload.allocation = {(uint32_t)(uintptr_t)tls, (uintptr_t)pc, (uintptr_t)addr, (uintptr_t)size};
                 write_log_sync(buf);
             }
 
             virtual void deallocate(tls_t tls, void* addr) {
                 ipc::event::BufferEntry buf{Type::FREE};
-                buf.payload.allocation = {(uint32_t)(uintptr_t)tls, (uint64_t)0x0, (uint64_t)addr, (uint64_t)0x0};
+                buf.payload.allocation = {(uint32_t)(uintptr_t)tls, (uintptr_t)0x0, (uintptr_t)addr, (uintptr_t)0x0};
                 write_log_sync(buf);
             }
 
             virtual void fork(tid_t parent, tid_t child, tls_t * tls) {
-                *tls = (void*)((uint64_t)child);
+                *tls = (void*)((uintptr_t)child);
                 ipc::event::BufferEntry buf{Type::FORK};
-                buf.payload.forkjoin = {(uint32_t)(uintptr_t)parent, (uint32_t)child};               
+                buf.payload.forkjoin = {(uint32_t)(uintptr_t)parent, (uint32_t)child};
                 write_log_sync(buf);
             }
