+++ drace-client/include/function-wrapper/internal.h
@@ -22,7 +22,7 @@ namespace drace {
 		using wrapcb_post_t = void(void *, void *);
 
 		namespace internal {
-			void wrap_dotnet_helper(uint64_t addr);
+			void wrap_dotnet_helper(uintptr_t addr);
 			bool wrap_function_clbck(const char *name, size_t modoffs, void *data);
 
 			bool mutex_wrap_callback(const char *name, size_t modoffs, void *data);
