+++ drace-client/src/symbol/Symbols.cpp
@@ -42,8 +42,8 @@ namespace symbol {
 
 		if (modptr) {
 			// Reverse search from pc until symbol can be decoded
-			uint64_t offset = pc - modptr->base;
-			auto limit = std::max((uint64_t)0, offset - (uint64_t)max_distance);
+			uintptr_t offset = pc - modptr->base;
+			auto limit = std::max((uintptr_t)0, offset - (uintptr_t)max_distance);
 			for (; offset >= limit; --offset) {
 				drsym_error_t err = drsym_lookup_address(modptr->info->full_path, offset, &syminfo, DRSYM_DEMANGLE);
 				if (err == DRSYM_SUCCESS || err == DRSYM_ERROR_LINE_NOT_AVAILABLE) {
@@ -73,8 +73,8 @@ namespace symbol {
 			sloc.mod_name = dr_module_preferred_name(modptr->info);
 
 			// Reverse search from pc until symbol can be decoded
-			uint64_t offset = pc - modptr->base;
-			auto limit = std::max((uint64_t)0, offset - (uint64_t)max_distance);
+			uintptr_t offset = pc - modptr->base;
+			auto limit = std::max((uintptr_t)0, offset - (uintptr_t)max_distance);
 			for (; offset >= limit; --offset) {
 				drsym_error_t err = drsym_lookup_address(modptr->info->full_path, offset, &syminfo, DRSYM_DEMANGLE);
 				if (err == DRSYM_SUCCESS || err == DRSYM_ERROR_LINE_NOT_AVAILABLE) {
